// src/types/vehiculoInterface.tsx
export interface Vehiculo {
  aduana?: any;
  agencia_distribuidora_del_vehiculo?: any;
  capacidad?: any;
  capacidad_de_pasajeros?: any;
  cilindraje?: any;
  clase_del_vehiculo?: 1;
  clave_de_la_secretaria_de_hacienda_y_credito_publico?: any;
  clave_de_nacional?: string;
  clave_de_taxi?: string;
  clave_del_documento_de_regularizacion?: any;
  clave_del_importador?: any;
  clave_municipal?: any;
  clave_vehicular?: any;
  color_del_vehiculo?: string;
  descripcion_de_la_agencia?: string;
  direccion?: any;
  empresa_propietaria?: any;
  es_de_servicio_publico?: boolean;
  es_foraneo?: boolean;
  es_propietario_ciudadano?: boolean;
  es_vehiculo_nacional?: boolean;
  estados_globales?: number;
  estatus_del_vehiculo?: number;
  exento_de_pago_de_tenencia?: boolean;
  fecha?: any;
  fecha_de_expiracion_de_la_placa?: any;
  fecha_de_la_factura?: any;
  fecha_de_pedimento?: any;
  fecha_de_repuve?: any;
  id?: number;
  importe_de_la_factura?: number;
  impuesto_del_anio_2016?: any;
  marca_vehicular?: any;
  modelo_del_vehiculo?: number;
  nofaca?: any;
  nombre_de_la_aduana?: any;
  numero_de_calcomania?: string;
  numero_de_chasis?: any;
  numero_de_cilindros?: number;
  numero_de_factura?: string;
  numero_de_motor?: string;
  numero_de_patente?: any;
  numero_de_pedimento?: any;
  numero_de_placa?: string;
  numero_de_puertas?: number;
  numero_de_registro_publico_vehicular?: any;
  observaciones?: string;
  pais?: number;
  peso_del_vehiculo?: any;
  plantilla_de_tramites?: any;
  propietario?: number;
  recargo_del_anio_2016?: any;
  registro_federal_vehicular?: any;
  registro_publico_vehicular?: string;
  serial?: string;
  serie?: string;
  servicio?: number;
  status?: boolean;
  tarjeta_de_circulacion?: any;
  tipo_de_combustible?: number;
  tipo_de_movimiento_repuve?: string;
  tipo_de_tencia_vehicular?: any;
  tipo_de_vehiculo?: number;
  usuario_verificador?: any;
  verificacion_tecnica_vehicular?: any;
  web_service?: boolean;
}